+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Converted `Compiler` from object to class
 - Added functions:
   - Conversion: `ToBoolean`, `ToComplex`, `ToInteger`, `ToMatrix`, `ToRational`, `ToSet`, `ToString`
+- Moved command information to library as JSON files
 - Fixed existing functions to support variables
 
 ### v0.3.3 (Feb 5, 2022)